Attempt2/requirements.txt
langchain>=0.2.0
langchain-huggingface>=0.1.2
transformers>=4.39.0
sentence-transformers>=2.6.0
torch>=2.1.2
streamlit>=1.29.0
python-dotenv>=1.0.0
google-api-python-client>=2.108.0
google-auth-oauthlib>=1.1.0
slack-sdk>=3.26.0
pandas>=2.1.3
sqlalchemy>=2.0.23
bitsandbytes>=0.41.1
accelerate>=0.25.0
